backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
